Fragile fix for a bug in las-norms due to overflow in cast from double to int

+++ sieve/las-norms.cpp
@@ -574,7 +574,7 @@ void lognorm_fill_rat_smart_inner (unsigned char *S, int i0, int i1, unsigned in
      */
 
     for (unsigned int j = j0; j < j1; j++) {
-	int i = i0;
+	int64_t i = i0;
 	double g = fabs(u0 * j + u1 * i0);
 	uint8_t y;
 	double root = -u0 * j / u1;
@@ -617,9 +617,9 @@ void lognorm_fill_rat_smart_inner (unsigned char *S, int i0, int i1, unsigned in
                 /* conversion rounding is towards zero, while we want it
                  * to be towards +infinity. */
                 ix += (ix >= 0);
-		if (UNLIKELY(ix >= i1))
+		if (UNLIKELY((int64_t)ix >= i1))
 		    ix = i1;
-		size_t di = (int) ix - i;	/* The cast matters ! */
+		size_t di = (int64_t) ix - i;	/* The cast matters ! */
 		i = ix;
 		if (!di)
                     break;
@@ -675,12 +675,12 @@ void lognorm_fill_rat_smart_inner (unsigned char *S, int i0, int i1, unsigned in
              * to be towards +infinity. */
             ix += (ix >= 0);
             // ASSERT ((int) ix >= i); // see bug 21518
-            if (LIKELY((int) ix >= i)) {
+            if (LIKELY((int64_t) ix >= i)) {
                 /* It is most likely that we'll only see this branch. Yet
                  * bug 21518 seems to trigger a nasty corner case */
                 if (UNLIKELY(ix >= i1))
                     ix = i1;
-                size_t di = (int) ix - i;	/* The cast matters ! */
+                size_t di = (int64_t) ix - i;	/* The cast matters ! */
                 i = ix;
                 memset_with_writeahead(S, y, di, MEMSET_MIN);
                 S += di;
